container-host-agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1
PyJWT[crypto]==2.8.0
docker==6.1.3
psutil==5.9.5
httpx==0.24.1
